requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-anthropic>=0.1.0
langgraph>=0.0.50
langgraph-supervisor>=0.0.1
pandas>=2.0.0
python-dotenv>=1.0.0
markdown>=3.4.0

